+++ types.go
@@ -264,76 +264,3 @@ func getTimeTicks[T any](val T) (int64, error) {
 	base := time.Date(1970, time.January, 1, ti.Hour(), ti.Minute(), ti.Second(), ti.Nanosecond(), time.UTC)
 	return base.UnixMicro(), err
 }
-
-func getMappedArrayValue[T any](lt mapping.LogicalType, val T) (*mapping.Value, error) {
-	var childValues []mapping.Value
-	childType := mapping.ArrayTypeChildType(lt)
-	defer mapping.DestroyLogicalType(&childType)
-
-	vSlice, err := toAnySlice(val)
-	if err != nil {
-		return nil, fmt.Errorf("could not cast %T to []any: %s", val, err)
-	}
-
-	for _, v := range vSlice {
-		vv, err := createValue(childType, v)
-		if err != nil {
-			return nil, fmt.Errorf("could not create value %s", err)
-		}
-		childValues = append(childValues, *vv)
-	}
-
-	arrValue := mapping.CreateArrayValue(childType, childValues)
-	return &arrValue, nil
-}
-
-func getMappedListValue[T any](lt mapping.LogicalType, val T) (*mapping.Value, error) {
-	var childValues []mapping.Value
-	childType := mapping.ListTypeChildType(lt)
-	defer mapping.DestroyLogicalType(&childType)
-
-	vSlice, err := toAnySlice(val)
-	if err != nil {
-		return nil, fmt.Errorf("could not cast %T to []any: %s", val, err)
-	}
-
-	for _, v := range vSlice {
-		vv, err := createValue(childType, v)
-		if err != nil {
-			return nil, fmt.Errorf("could not create value %s", err)
-		}
-		childValues = append(childValues, *vv)
-	}
-
-	listValue := mapping.CreateListValue(childType, childValues)
-	return &listValue, nil
-}
-
-func getMappedStructValue(lt mapping.LogicalType, val any) (*mapping.Value, error) {
-	var values []mapping.Value
-	vMap, ok := val.(map[string]any)
-	if !ok {
-		return nil, fmt.Errorf("could not cast %T to map[string]any", val)
-	}
-
-	childCount := mapping.StructTypeChildCount(lt)
-
-	for i := mapping.IdxT(0); i < childCount; i++ {
-		childName := mapping.StructTypeChildName(lt, i)
-		childType := mapping.StructTypeChildType(lt, i)
-		defer mapping.DestroyLogicalType(&childType)
-		v, exists := vMap[childName]
-		if exists {
-			vv, err := createValue(childType, v)
-			if err != nil {
-				return nil, fmt.Errorf("could not create value %s", err)
-			}
-			values = append(values, *vv)
-		} else {
-			values = append(values, mapping.CreateNullValue())
-		}
-	}
-
-	structValue := mapping.CreateStructValue(lt, values)
-	return &structValue, nil
-}
